pyproject.toml
[tool.poetry]
name = "rule_enginee"
version = "0.1.0"
description = ""
authors = ["RahulDas-dev <r.das699@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
xmltodict = "^0.13.0"
jsonata-python = "^0.4.0"
litellm = "^1.47.1"
anthropic = {extras = ["vertex"], version = "^0.34.2"}
google-cloud-aiplatform = "^1.67.1"
termcolor = "^2.4.0"
sqlalchemy = "^2.0.35"
loadenv = "^0.1.1"
tenacity = "^9.0.0"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
ujson = "^5.10.0"
sqlalchemy-utils = "^0.41.2"
python-multipart = "^0.0.10"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
pyclean = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".ruff_cache",
    ".git",
    ".ipynb_checkpoints",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "notebooks",
    "aider"
]
extend-include = []
line-length = 120
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
ignore = ["ANN204", "ANN401", "E731", "D", "DTZ005", "BLE001","B008", "CPY001","COM812", "ERA001", "EM101", "FA","FBT", "G004", "UP", "TRY", "PTH123","ISC001" ]
select = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
